Pin the lint rule set so CI is reproducible, and clear the findings it keeps
CI failed on 100 lint errors, most of them in files this branch never touched:
ruff's *default* rule selection broadened in a newer release, so an unpinned
linter turned the branch red without a line of code changing. main would fail
the same way today.

Fix the cause rather than the symptom: declare the rule set and target version
explicitly in pyproject, so the linter's opinion is a property of the repo and
not of whichever ruff the runner happened to install. The selection adds I, B,
RUF and BLE to the classic defaults, and ignores RUF002/RUF003 because the
docstrings here use en dashes and maths symbols deliberately.

Then clear every finding under that set:

* lazy optional-backend imports (OpenCV, InsightFace) keep their blind except
  with an explicit noqa and a reason — falling back on *any* import failure is
  the intent
* replace a mutable-default cache in the sensor detector with lru_cache
* drop redundant int(round(...)) casts, unpack only what is used, and use
  unpacking over list concatenation
* sort imports and __all__

+++ src/faceguard/liveness/parallax.py
@@ -106,7 +106,7 @@ def _widest_baseline(gray, blocks, probes: int = 4):
     # Average a few adjacent frames per endpoint: cuts noise without materially
     # changing the pose at either end of the baseline.
     span = max(1, t // 12)
-    marks = [int(round(f * (t - span))) for f in np.linspace(0.0, 1.0, probes + 1)]
+    marks = [round(f * (t - span)) for f in np.linspace(0.0, 1.0, probes + 1)]
     ends = [np.mean(gray[m:m + span], axis=0) for m in marks]
 
     best_disp, best_motion = None, -1.0
